+++ packages/react/src/UpSetJS.tsx
@@ -9,16 +9,17 @@ import React, { useMemo, forwardRef, Ref, useCallback } from 'react';
 import { UpSetProps } from './interfaces';
 import { exportSVG, exportVegaLite } from './exporter';
 import { exportDump, exportSharedLink } from './exporter/exportDump';
-import deriveDataDependent from './components/deriveDataDependent';
-import deriveSizeDependent from './components/deriveSizeDependent';
-import deriveStyleDependent from './components/deriveStyleDependent';
+import deriveDataDependent from './derive/deriveDataDependent';
+import deriveSizeDependent from './derive/deriveSizeDependent';
+import deriveStyleDependent from './derive/deriveStyleDependent';
 import ExportButtons from './components/ExportButtons';
 import QueryLegend from './components/QueryLegend';
 import UpSetAxis from './components/UpSetAxis';
 import UpSetChart from './components/UpSetChart';
 import UpSetQueries from './components/UpSetQueries';
 import UpSetSelection from './components/UpSetSelection';
-import { clsx, generateId } from './components/utils';
+import { clsx } from './components/utils';
+import { generateId } from './derive/utils';
 import { fillDefaults } from './fillDefaults';
 import { baseRules } from './rules';
