+++ packages/react/src/components/CombinationSelectionChart.tsx
@@ -8,9 +8,9 @@
 import { ISetLike, ISetCombination } from '@upsetjs/model';
 import React, { PropsWithChildren, ReactNode } from 'react';
 import { UpSetAddons } from '../interfaces';
-import { UpSetDataInfo } from './deriveDataDependent';
-import { UpSetSizeInfo } from './deriveSizeDependent';
-import { UpSetStyleInfo } from './deriveStyleDependent';
+import { UpSetDataInfo } from '../derive/deriveDataDependent';
+import { UpSetSizeInfo } from '../derive/deriveSizeDependent';
+import { UpSetStyleInfo } from '../derive/deriveStyleDependent';
 import { addonPositionGenerator, clsx } from './utils';
 
 function CombinationSelectionChart<T>({
