+++ packages/react/src/components/UpSetQueries.tsx
@@ -9,9 +9,9 @@ import { ISetLike, queryElemOverlap, queryOverlap, UpSetQuery } from '@upsetjs/m
 import React, { PropsWithChildren, useMemo } from 'react';
 import { UpSetAddon, UpSetAddonProps } from '../interfaces';
 import CombinationSelectionChart from './CombinationSelectionChart';
-import { UpSetDataInfo } from './deriveDataDependent';
-import { UpSetSizeInfo } from './deriveSizeDependent';
-import { UpSetStyleInfo } from './deriveStyleDependent';
+import { UpSetDataInfo } from '../derive/deriveDataDependent';
+import { UpSetSizeInfo } from '../derive/deriveSizeDependent';
+import { UpSetStyleInfo } from '../derive/deriveStyleDependent';
 import SetSelectionChart from './SetSelectionChart';
 
 const EMPTY_ARRAY: any[] = [];
